- Moved the types the user must use to a separate DLL. - Added more redundant file extensions to the .gitignore file.
