+++ smg/smg.common/Properties/AssemblyInfo.cs
@@ -1,14 +1,15 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("smg")]
+[assembly: AssemblyTitle("smg.UserTypes")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("smg")]
+[assembly: AssemblyProduct("smg.UserTypes")]
 [assembly: AssemblyCopyright("Copyright ©  2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -19,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("ae9d5bef-cc70-466f-b1f9-965e7acb7fd7")]
+[assembly: Guid("27c2ed74-99b8-4b26-83d1-807b375629cc")]
 
 // Version information for an assembly consists of the following four values:
 //
