+++ smg/smg.common/StateDescription/Attributes/AvailableForAllStatesAttribute.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace smg.StateDescription.Attributes
+namespace smg.common.StateDescription.Attributes
 {
     /// <summary>
     /// Indicates that a method is available for all of the states that the stateful object containing it might be at.
