+++ smg/smg.common/StateDescription/Attributes/StateGroupAttribute.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace smg.StateDescription.Attributes
+namespace smg.common.StateDescription.Attributes
 {
     /// <summary>
     /// Represents a group of states where only one of the states can be active.
