+++ smg/smg/ArgumentParsing/ArgsParser.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Security.Cryptography.X509Certificates;
 using smg.ArgumentParsing.Attributes;
 using smg.ArgumentParsing.Exceptions;
