+++ smg/smg/ExtensionMethods/EnumerableExtensions.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace smg.ExtensionMethods
 {
