+++ smg/smg/StateGeneration/ExtensionMethods/TypeExtensions.cs
@@ -3,8 +3,6 @@ using System.CodeDom;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace smg.StateGeneration.ExtensionMethods
 {
