+++ smg/smg/StateMachineGenerator.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 using System.CodeDom;
 using System.CodeDom.Compiler;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 using smg.ArgumentParsing;
 using smg.StateGeneration;
